数值系统基础款完成 但暂未考虑建筑物类型
下面，完成建筑物级别抽象
给角色添加多种攻击盒子
修复角色旋转，攻击盒无法旋转bug

+++ Scripts/Entity/Core/AttackSystem/Core/IDamageable.cs
@@ -9,6 +9,6 @@ namespace DDemo.Scripts.Entity.Core.AttackSystem.Core
 {
     public interface IDamageable
     {
-        void TakeDamage(Node2D attacker,int attackValue);
+        void TakeDamage(Node2D attacker,float attackValue);
     }
 }
